+++ src/Hacks/Bypass.cpp
@@ -9,12 +9,14 @@ using namespace geode::prelude;
 // Icon Bypass
 class $modify(GameManager) {
     bool isIconUnlocked(int _id, IconType _type) {
-        if (!Hacks::isHackEnabled("Icon Bypass") || Hacks::isHackEnabled("Enable Patching")) return GameManager::isIconUnlocked(_id, _type);
-        return true;
+        if (GameManager::isIconUnlocked(_id, _type)) return true;
+        if (_id <= 0) return false;
+        //if (_id >= GameManager::countForType(_type)) return false;
+        return Hacks::isHackEnabled("Icon Bypass");
     }
     bool isColorUnlocked(int _id, UnlockType _type) {
-        if (!Hacks::isHackEnabled("Icon Bypass") || Hacks::isHackEnabled("Enable Patching")) return GameManager::isColorUnlocked(_id, _type);
-        return true;
+        if (GameManager::isColorUnlocked(_id, _type)) return true;
+        return Hacks::isHackEnabled("Icon Bypass");
     }
 };